+++ lib/assets/javascripts/new-dashboard/components/Catalog/NotAvailable.vue
@@ -53,7 +53,7 @@ export default {
     },
     description: {
       type: String,
-      required: true
+      required: false
     },
     contactUrl: {
       type: String,